+++ CHANGELOG.md
@@ -37,6 +37,7 @@ Per Keep a Changelog there are 6 main categories of changes:
   - Added function for writing previous wgpu textures with new rgba data and exporting to egui::TextureId @AlbinSjoegren
 - rend3-routine: Added the option to set a custom primitive topology value when building a forward routine. @setzer22
 - rend3-routine: Added a resolution field to the per-frame uniforms. @setzer22
+- rend3-routine: Added add_clear_to_graph to make clears explicit and add `clear_color` argument to base rendergraph.
 
 ### Fixes
 - Fixed mismatched BGLs when using a custom material with no cutout specification